refactor: rename ShaderMaterialNames file - more accurate filename to the contents of the file
