+++ setup_wizard/domain/shader_material_names.py
