+++ setup_wizard/material_data_import_setup/material_data_applier.py
@@ -8,7 +8,7 @@ from bpy.types import Material
 from setup_wizard.domain.character_types import CharacterType
 from setup_wizard.domain.game_types import GameType
 from setup_wizard.domain.outline_material_data import OutlineMaterialGroup
-from setup_wizard.domain.shader_materials import V3_BonnyFestivityGenshinImpactMaterialNames
+from setup_wizard.domain.shader_material_names import V3_BonnyFestivityGenshinImpactMaterialNames
 
 
 class MaterialDataAppliersFactory:
